+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 - New option `--no-v-prefix` to generate the tag names without prepending `v` [#43]
 - New option `--sortReleases` to enable or disable the sorting of releases [#51]
+- `--help, -h` flag [#45].
 
 ### Fixed
 - Updated dependencies
@@ -102,6 +103,7 @@ New version merging Deno and Node code using Deno's `dnt` package.
 [#41]: https://github.com/oscarotero/keep-a-changelog/issues/41
 [#42]: https://github.com/oscarotero/keep-a-changelog/issues/42
 [#43]: https://github.com/oscarotero/keep-a-changelog/issues/43
+[#45]: https://github.com/oscarotero/keep-a-changelog/issues/45
 [#51]: https://github.com/oscarotero/keep-a-changelog/issues/51
 
 [2.6.0]: https://github.com/oscarotero/keep-a-changelog/compare/v2.5.3...v2.6.0